Move addon from pods structure to default structure
