+++ addon/components/validated-form.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import layout from './template';
+import layout from '../templates/components/validated-form';
 
 function runTaskOrAction(taskOrAction, model) {
   return taskOrAction.perform ?
