+++ app/components/validated-button.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-validated-form/components/validated-button';
