--- app/components/validated-button/component.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-validated-form/components/validated-button/component';
