+++ app/components/validated-form.js
@@ -1,0 +1,6 @@
+import ValidatedForm from 'ember-validated-form/components/validated-form';
+import config from '../config/environment';
+
+export default ValidatedForm.extend({
+  config: config['ember-validated-form']
+});
