+++ app/components/validated-input.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-validated-form/components/validated-input';
