--- app/components/validated-input/component.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-validated-form/components/validated-input/component';
